:art: refactor : 입력 받은 데이터 검증 로직을 도메인으로 추출
