+++ src/main/java/ladder/step2/controller/LadderGame.java
@@ -4,13 +4,14 @@ import ladder.step2.domain.Ladder;
 import ladder.step2.domain.LadderFactory;
 import ladder.step2.domain.Players;
 import ladder.step2.domain.strategy.RandomLineCreateStrategy;
+import ladder.step2.dto.LadderHeightDTO;
 import ladder.step2.view.input.InputView;
 import ladder.step2.view.output.ResultView;
 
 public class LadderGame {
     public void play() {
         Players players = InputView.inputPlayerNames();
-        int ladderHeight = InputView.inputLadderHeight();
+        LadderHeightDTO ladderHeight = InputView.inputLadderHeight();
         
         Ladder ladder = LadderFactory.of(players.size(), ladderHeight, new RandomLineCreateStrategy());
         resultPrint(players, ladder);
