+++ src/test/java/ladder/step2fixture/domain/PlayersFixture.java
@@ -2,8 +2,6 @@ package ladder.step2fixture.domain;
 
 import ladder.step2.domain.Players;
 
-import java.util.Arrays;
-
 public class PlayersFixture {
-    public static final Players PLAYERS = new Players(Arrays.asList(PlayerNameFixture.POBI, PlayerNameFixture.HONUX, PlayerNameFixture.JUN, PlayerNameFixture.JK));
+    public static final Players PLAYERS = new Players("pobi, honux, jun, jk");
 }
